Addition of allowing a self resolution.
If you try to do something like this on a class with an interface it won't resolve well.

BackgroundJob.Enqueue(() => this.SendSms(message));

The activator will try to instantiate the object, it will fail because it was never registered and then try to instantiate it by the first interface.

Example:

public class NavService : INavService {
  // do stuff
  public void SendStuff() {

  }
  pubic void Run() {
    BackgroundJob.Enqueue(() => this.SendStuff());
  }
}

container.Register<INavService, NavService>(new PerScopeLifetime());